requirements.txt
uvicorn
fastapi
google-generativeai
openai